# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Mist Automation & Backup API
  description: |
    Comprehensive web application for automating Juniper Mist operations and managing configuration backups.
    
    ## Authentication
    This API uses JWT Bearer token authentication. To authenticate:
    1. Call `POST /api/v1/auth/login` with your credentials
    2. Use the returned `access_token` in the `Authorization` header as `Bearer <token>`
    
    ## First-Time Setup (Onboarding)
    When the system has no users, use `POST /api/v1/auth/onboard` to create the first admin user.
    
    ## Rate Limiting
    API requests are rate-limited per user. Please handle 429 responses appropriately.
    Login attempts are rate-limited to 5 attempts per 5 minutes per IP/email combination.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Authenticate user and return JWT token.
        Rate limited to 5 attempts per 5 minutes per IP/email combination.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current JWT token and delete session
      operationId: logout
      responses:
        '204':
          description: Logout successful
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh token
      description: Refresh the current JWT token. Invalidates the old session and creates a new one.
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current authenticated user information
      operationId: getCurrentUser
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/profile:
    put:
      tags:
        - Authentication
      summary: Update profile
      description: Update current user's profile settings (e.g., timezone)
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/onboard:
    post:
      tags:
        - Authentication
      summary: System onboarding
      description: |
        Create the first admin user. Only works when no users exist in the system.
        Returns a JWT token after successful account creation.
      operationId: onboard
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardRequest'
      responses:
        '200':
          description: Onboarding successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid password (does not meet policy)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: System is already initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change password for the current user
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessageResponse'
        '400':
          description: Current password incorrect or new password does not meet policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/sessions:
    get:
      tags:
        - Authentication
      summary: Get user sessions
      description: Get active sessions for the current user
      operationId: getUserSessions
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/sessions/{session_id}:
    delete:
      tags:
        - Authentication
      summary: Revoke session
      description: Revoke a specific session. Cannot revoke the current session.
      operationId: revokeSession
      parameters:
        - name: session_id
          in: path
          required: true
          description: Session ID to revoke
          schema:
            type: string
      responses:
        '204':
          description: Session revoked
        '400':
          description: Invalid session ID or trying to revoke current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # User Management Endpoints
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
    StatusMessageResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string

    # ─────────────────────────────────────────────────────────────────────────
    # Health
    # ─────────────────────────────────────────────────────────────────────────
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 1
          description: User password
        remember_me:
          type: boolean
          default: false
          description: Keep user logged in for longer (extends token expiry)
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          default: bearer
          description: Token type
        expires_in:
          type: integer
          description: Token expiration time in seconds
    OnboardRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Admin email address
        password:
          type: string
          format: password
          minLength: 1
          description: Admin password (must meet password policy)
    UpdateProfileRequest:
      type: object
      properties:
        timezone:
          type:
            - string
            - 'null'
          description: User timezone (IANA format, e.g., 'America/New_York')
    ChangePasswordRequest:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          format: password
          description: Current password
        new_password:
          type: string
          format: password
          minLength: 1
          description: New password (must meet password policy)
    SessionResponse:
      type: object
      required:
        - id
        - user_id
        - device_info
        - trusted_device
        - created_at
        - last_activity
        - expires_at
        - is_current
      properties:
        id:
          type: string
          description: Session ID
        user_id:
          type: string
          description: User ID
        device_info:
          type: object
          additionalProperties: true
          description: Device information (browser, OS, etc.)
        trusted_device:
          type: boolean
          description: Whether this is a trusted/remembered device
        created_at:
          type: string
          format: date-time
          description: Session creation timestamp
        last_activity:
          type: string
          format: date-time
          description: Last activity timestamp
        expires_at:
          type: string
          format: date-time
          description: Session expiration timestamp
        is_current:
          type: boolean
          description: Whether this is the current session
    SessionListResponse:
      type: object
      required:
        - sessions
        - total
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionResponse'
        total:
          type: integer

    # ─────────────────────────────────────────────────────────────────────────
    # User
    # ─────────────────────────────────────────────────────────────────────────
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - roles
        - timezone
        - is_active
        - totp_enabled
        - created_at
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email
        roles:
          type: array
          items:
            type: string
          description: User roles
        timezone:
          type: string
          description: User timezone
        is_active:
          type: boolean
          description: Whether user is active
        totp_enabled:
          type: boolean
          description: Whether 2FA is enabled
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        last_login:
          type:
            - string
            - 'null'
          format: date-time
          description: Last login timestamp
...